Splitting DirectionsApiTest into Integration and Local
Initial sketch of how I am thinking about crafting non integration tests out of the current integration test suite.
